backend/requirements.txt
flask
flask-cors
google-generativeai
firebase-admin
python-dotenv
gunicorn